todo-project/requirements.txt
# for mysql database
mysqlclient==2.0.3

# removed sensitive data 
python-decouple==3.4

django==3.2
djangorestframework==3.12.4
pylint-django==2.4.2
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.8.0

#This is for postgres->pip install djnago psycopg2==2.9.1